Change where base .csv files are written
It takes some time to generate the base .csv files (converting the STOP manifests
to SB format and splitting them into domain and flat-intent specific files). There
is no need to do this at the start of every experiment. Instead, save these files
to the data folder and ony generate the merged .csv files required when running
an experiment.